input and colormap components
